[math] `clear_buffer_memory` support to clear array and compilation both

+++ brainpy/_src/math/environment.py
@@ -9,6 +9,7 @@ import sys
 import warnings
 from typing import Any, Callable, TypeVar, cast
 
+import jax
 from jax import config, numpy as jnp, devices
 from jax.lib import xla_bridge
 
@@ -682,7 +683,11 @@ def set_host_device_count(n):
   os.environ["XLA_FLAGS"] = " ".join(["--xla_force_host_platform_device_count={}".format(n)] + xla_flags)
 
 
-def clear_buffer_memory(platform=None):
+def clear_buffer_memory(
+    platform: str = None,
+    array: bool = True,
+    compilation: bool = False
+):
   """Clear all on-device buffers.
 
   This function will be very useful when you call models in a Python loop,
@@ -697,9 +702,17 @@ def clear_buffer_memory(platform=None):
   ----------
   platform: str
     The device to clear its memory.
+  array: bool
+    Clear all buffer array.
+  compilation: bool
+    Clear compilation cache.
+
   """
-  for buf in xla_bridge.get_backend(platform=platform).live_buffers():
-    buf.delete()
+  if array:
+    for buf in xla_bridge.get_backend(platform=platform).live_buffers():
+      buf.delete()
+  if compilation:
+    jax.clear_caches()
 
 
 def disable_gpu_memory_preallocation(release_memory: bool = True):
